+++ src/flexi_error.rs
@@ -91,6 +91,12 @@ impl From<glob::PatternError> for FlexiLoggerError {
         FlexiLoggerError::Io(std::io::Error::new(std::io::ErrorKind::Other, err))
     }
 }
+#[cfg(feature = "ziplogs")]
+impl From<zip::result::ZipError> for FlexiLoggerError {
+    fn from(err: zip::result::ZipError) -> FlexiLoggerError {
+        FlexiLoggerError::Io(std::io::Error::new(std::io::ErrorKind::Other, err))
+    }
+}
 
 #[cfg(feature = "specfile")]
 impl From<toml::de::Error> for FlexiLoggerError {